src/test/java/StatsServiceTest.java
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;
import ru.netology.stats.StatsService;

class StatsServiceTest {

    @Test
    void shouldFindMinSales() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 9;

        int actual = service.minSales(sales);

        assertEquals(expected, actual);

    }

    @Test
    void shouldGetSumOfAllSales() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 180;

        int actual = service.sumOfAllSales(sales);

        assertEquals(expected, actual);
    }

    @Test
    void shouldGetAverageSale() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 15;

        int actual = service.averageSale(sales);

        assertEquals(expected, actual);
    }

    @Test
    void shouldGetMonthOfMaxSales() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 8;

        int actual = service.monthOfMaxSales(sales);

        assertEquals(expected, actual);
    }

    @Test
    void shouldGetMonthsOfSalesBelowAverage() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;

        int actual = service.monthsOfSalesBelowAverage(sales);

        assertEquals(expected, actual);
    }

    @Test
    void shouldGetMonthsOfSalesAboveAverage() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;

        int actual = service.monthsOfSalesAboveAverage(sales);

        assertEquals(expected, actual);
    }
}